+++ warp/builtins.py
@@ -4098,6 +4098,14 @@ add_builtin(
     doc="Modulo operation using truncated division.",
     group="Operators",
 )
+add_builtin(
+    "mod",
+    input_types={"a": vector(length=Any, dtype=Scalar), "b": vector(length=Any, dtype=Scalar)},
+    constraint=sametypes,
+    value_func=sametypes_create_value_func(Scalar),
+    doc="Modulo operation using truncated division.",
+    group="Operators",
+)
 
 add_builtin(
     "div",