+++ warp/stubs.py
@@ -1813,6 +1813,12 @@ def mod(a: Scalar, b: Scalar) -> Scalar:
 
 
 @over
+def mod(a: Vector[Any, Scalar], b: Vector[Any, Scalar]) -> Scalar:
+    """Modulo operation using truncated division."""
+    ...
+
+
+@over
 def div(a: Scalar, b: Scalar) -> Scalar:
     """ """
     ...