+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add support for stream priorities to hint to the device that it should process pending work
   in high-priority streams over pending work in low-priority streams when possible
   ([docs](https://nvidia.github.io/warp/modules/concurrency.html#stream-priorities)).
+- Add `wp.mod()` for vector types ([GH-282](https://github.com/NVIDIA/warp/issues/282)).
 
 ## [1.3.1] - 2024-07-27